+++ commerce/payment/src/main/resources/schema.sql
@@ -2,9 +2,9 @@ create table if not exists payments
 (
     payment_id     uuid default gen_random_uuid() primary key,
     order_id       uuid not null,
-    products_total double precision,
-    delivery_total double precision,
-    total_payment  double precision,
-    fee_total      double precision,
+    products_total numeric(19,4),
+    delivery_total numeric(19,4),
+    total_payment  numeric(19,4),
+    fee_total      numeric(19,4),
     status         varchar(15)
 );
